kbuild: clang: Disable the 'duplicate-decl-specifier' warning
clang generates plenty of these warnings in different parts of the code.
They are mostly caused by container_of() and other macros which declare
a "const <type> *" variable for their internal use which triggers a
"duplicate 'const' specifier" warning if the <type> is already const
qualified.

For the above cases the warnings are only raised when -std=gnu89 is set,
however this doesn't trigger the warning with gcc. The longer term
solution is probably to make clang behave like gcc in this aspect, in
the meantime disable the warning.

BUG=chromium:702741,chromium:723720
TEST=build for gru and pyro with clang

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: Ie922580dbdaf61f03843411914b6b480d319576a
Reviewed-on: https://chromium-review.googlesource.com/506553
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>